requirements/base.txt
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
setuptools==70.0.0
psycopg2==2.9.9
argon2-cffi==23.1.0
pillow==10.3.0
requests==2.32.3
pandas==2.2.2
lxml==5.2.2
beautifulsoup4-4.12.3